pyproject.toml
[tool.poetry]
name = "SHIFT_TZ"
version = "0.1.0"
description = ""
authors = ["Nikita Makoveev <nikitamakoveev564@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pytest = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
